+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2020-08-27 - Fixed bug in image area feature if link attribute was left empty.
 * 2020-08-27 - Fixed regression for commit Added rule to show description border for all users for restricted activities.
 
 ### Release v3.8-r3